created utility for todo item functions
